Revert "chore: moved repo operations to own file"
This reverts commit d6d5e914339c0da541da6415b459dca26723869b.
